pyproject.toml
[tool.poetry]
name = "job-search-v2"
version = "0.1.0"
description = ""
authors = ["Thibaut Lemarchand <git@tlemarchand.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
schedule = "^1.2.2"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
pytest-flask = "^1.3.0"
jupyter = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
